common.scm, fsm.scm: Remove dependency on 'guile-lib'
Some modules use 'define-with-docs' which is part of 'guile-lib'
package.  The dependency is not documented and leads to errors when
the 'guile-lib' is missing.

This patch removes aforementioned dependency as it seems not worthy of
requiring the whole library only just for 'define-with-docs'
facilities.

Reported by Alex Sassmannshausen <alex.sassmannshausen@gmail.com>

* modules/ics/common.scm, modules/ics/fsm.scm: Don't use
  'define-with-docs'.