modules/ics/fsm/content-line-context: Bugfix
Guile-ICS would throw an error when handling multiple instances of the same
property parameter in one content line.  According to RFC 6350 multiple
instances of the same parameter in a content line should be read as multiple
values for the same parameter (in the same way as the RFC specifies the value
lists.)

This patch should fix this error.

Reported by Felix Lechner in
<https://github.com/artyom-poptsov/guile-ics/issues/1>

* modules/ics/fsm/content-line-context.scm: (content-line:store-param-value):
  Consider multiple instances of the same parameter as additional values for the
  parameter (as in a list) instead of throwing an error.
* tests/content-line-parser.scm ("multiple instances of a parameter"): New test.